requirements.txt
antlr4-python3-runtime==4.8
feather-format==0.4.1
gitdb==4.0.5
GitPython==3.1.13
hydra-core==1.0.6
joblib==1.0.1
numpy==1.20.1
omegaconf==2.0.6
pandas==1.2.2
pyarrow==3.0.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
smmap==3.0.5
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
